Buttons for Edit/Delete are now visible only to a specific user role

+++ OpenDiscussion/Controllers/CategoriesController.cs
@@ -20,13 +20,22 @@ namespace OpenDiscussion.Controllers
             // var categories = db.Categories.Include("categories");
             var categories = from category in db.Categories select category;
             ViewBag.Categories = categories;
+            SetAccessRights();
             return View();
         }
         [Authorize(Roles = "Admin")]
         public IActionResult Edit(int id)
         {
-            Category category = db.Categories.Find(id);
-            return View(category);
+            if (User.IsInRole("Admin"))
+            {
+                Category category = db.Categories.Find(id);
+                return View(category);
+            }
+            else
+            {
+                TempData["Message"] = "Nu aveti voie sa editati o categorie";
+                return RedirectToAction("Index");
+            }
         }
         [Authorize(Roles = "Admin")]
         [HttpPost]
@@ -36,20 +45,29 @@ namespace OpenDiscussion.Controllers
             requestCategory.CategoryId = id;
             if (ModelState.IsValid)
             {
-                category.Name = requestCategory.Name;
-                category.Description = requestCategory.Description;
-                db.SaveChanges();
+                if (User.IsInRole("Admin"))
+                {
+                    category.Name = requestCategory.Name;
+                    category.Description = requestCategory.Description;
+                    db.SaveChanges();
+                }
+                else
+                    TempData["Message"] = "Nu aveti voie sa editati o categorie";
                 return RedirectToAction("Index");
             }
             else
-            {
                 return View(requestCategory);
-            }
         }
         [Authorize(Roles = "Admin")]
         public IActionResult New()
         {
-            return View();
+            if(User.IsInRole("Admin"))
+                return View();
+            else
+            {
+                TempData["Message"] = "Nu aveti voie sa adaugati o categorie";
+                return RedirectToAction("Index");
+            }
         }
         [Authorize(Roles = "Admin")]
         [HttpPost] 
@@ -57,8 +75,13 @@ namespace OpenDiscussion.Controllers
         {
             if (ModelState.IsValid)
             {
-                db.Categories.Add(requestCategory);
-                db.SaveChanges();
+                if (User.IsInRole("Admin"))
+                {
+                    db.Categories.Add(requestCategory);
+                    db.SaveChanges();
+                }
+                else
+                    TempData["Message"] = "Nu aveti voie sa adaugati o categorie";
                 return RedirectToAction("Index");
             }
             else
@@ -68,10 +91,21 @@ namespace OpenDiscussion.Controllers
         [HttpPost]
         public IActionResult Delete(int id)
         {
-            Category category = db.Categories.Include("Topics").Where(cat => cat.CategoryId == id).First();
-            db.Categories.Remove(category);
-            db.SaveChanges();
+            if (User.IsInRole("Admin"))
+            {
+                Category category = db.Categories.Include("Topics").Where(cat => cat.CategoryId == id).First();
+                db.Categories.Remove(category);
+                db.SaveChanges();
+            }
+            else
+                TempData["Message"] = "Nu aveti voie sa stergeti o categorie";
             return RedirectToAction("Index");
+        }
+        private void SetAccessRights()
+        {
+            ViewBag.ShowButtons = false;
+            if(User.IsInRole("Admin"))
+                ViewBag.ShowButtons = true;
         }
     }
 }
